Consolidate duplicated helpers into shared detail headers
FAB/plotfile header parsing (four parseIntegers copies, two box parsers, two
descriptor parsers, five ghost-grow variants), the query layer's block-lookup
helpers (verified byte-identical between SliceQuery and LineQuery), scalar
plane validation (three drifting variants), and printf-format scanning each
get one definition. Every drift resolves toward the strict variant: catalog
descriptors now validate byte order up front, DatasetExtract uses the
overflow-checked valueOffset, VectorGlyphs rejects negative dimensions, and
conversionSpecifier checks the conversion character. Throwing io helpers are
templated on the reader's error type so no thrown types change.
